LPS-61709 Improve error handling

+++ modules/apps/configuration-admin/configuration-admin-web/src/main/java/com/liferay/configuration/admin/web/portlet/action/DeleteConfigurationMVCActionCommand.java
@@ -60,6 +60,10 @@ public class DeleteConfigurationMVCActionCommand implements MVCActionCommand {
 			Configuration configuration =
 				_configurationModelRetriever.getConfiguration(pid);
 
+			if (configuration == null) {
+				return false;
+			}
+
 			configuration.delete();
 		}
 		catch (IOException ioe) {
